Naming convention again

+++ DotEngine/Engine/QuadTree.cpp
@@ -37,7 +37,7 @@ bool QuadTree::Insert(PhysicsComponent* newNode)
         SubDivide();
     }
     for(QuadTree* tree : m_directions){
-        for(unsigned int i = 0; i < m_nodes.size(); i++){
+        for(size_t i = 0; i < m_nodes.size(); i++){
             tree->Insert(m_nodes[i]);
             // if(!nodes[i]) continue;
             // if(tree->Insert(nodes[i])){
